update pow evmos
