+++ src/routes/safe/components/Balances/Coins/styles.ts
@@ -1,7 +1,8 @@
 import { sm, bgBox } from 'src/theme/variables'
 import { createStyles } from '@material-ui/core'
-
-export const styles = createStyles({
+import styled from 'styled-components'
+import { Button } from '@gnosis.pm/safe-react-components'
+const styles = createStyles({
   iconSmall: {
     fontSize: 16,
   },
@@ -32,3 +33,15 @@ export const styles = createStyles({
     backgroundColor: bgBox,
   },
 })
+
+const StyledButton = styled(Button)`
+  &&.MuiButton-root {
+    margin: 4px 12px 4px 0px;
+    padding: 0 12px;
+    min-width: auto;
+  }
+  svg {
+    margin: 0 6px 0 0;
+  }
+`
+export { StyledButton, styles }
